# specs/001-interest-media-sharing/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Interest-Centred Media Sharing API
  version: 0.1.0
  description: >
    HTTP contract for the interest-centred media sharing app.
    Derived from specs/001-interest-media-sharing/spec.md; every operation cites the
    functional requirements it serves.

    Visibility note: every response that can contain posts has already passed through
    the single visibility boundary described in research.md D6. A caller never receives
    a post excluded by FR-014 to FR-016, on any endpoint (FR-018, SC-009).
servers:
  - url: https://api.example.com/v1
security:
  - bearerAuth: []
...
paths:
  /me:
    get:
      tags: [People]
      summary: Current person's profile and preferences
      description: Serves FR-002, FR-049.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SelfProfile' }
        '401': { $ref: '#/components/responses/Unauthorised' }
    patch:
      tags: [People]
      summary: Update display name, avatar, bio, or notification preferences
      description: Serves FR-002, FR-049.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProfileUpdate' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SelfProfile' }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      tags: [People]
      summary: Delete the current account
      description: >
        Serves FR-003. Returns immediately with the retention outcome; content removal
        and comment anonymisation proceed asynchronously. Followers-only content becomes
        inaccessible as soon as this returns.
      responses:
        '202':
          description: Deletion accepted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeletionOutcome' }
  /people/{handle}:
    get:
      tags: [People]
      summary: A person's public profile
      description: >
        Serves FR-038. Includes follower and following counts and the interests they
        post to most. Posts are fetched separately via /people/{handle}/posts.
      parameters:
        - $ref: '#/components/parameters/Handle'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PublicProfile' }
        '404': { $ref: '#/components/responses/NotFound' }
  /people/{handle}/posts:
    get:
      tags: [People, Posts]
      summary: A person's posts, newest first
      description: Serves FR-038. Filtered to what the caller may see.
      parameters:
        - $ref: '#/components/parameters/Handle'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PostPage' }
  /people/{handle}/follow:
    put:
      tags: [People]
      summary: Follow a person
      description: >
        Serves FR-037. Idempotent. Following a person does not by itself add their
        posts to the caller's home feed — see FR-033 and GET /feed/home.
      parameters:
        - $ref: '#/components/parameters/Handle'
      responses:
        '204': { description: Following }
        '409':
          description: Blocked in either direction (FR-044)
    delete:
      tags: [People]
      summary: Unfollow a person
      description: Serves FR-037. Idempotent.
      parameters:
        - $ref: '#/components/parameters/Handle'
      responses:
        '204': { description: Not following }
  /interests:
    get:
      tags: [Interests]
      summary: Browse or search interests
      description: >
        Serves FR-025, FR-026. With `q`, performs type-ahead matching across both
        levels, returning each sub-interest with its parent. With `level=top` and no
        `q`, returns the curated catalogue.
      security: []
      parameters:
        - name: q
          in: query
          schema: { type: string, minLength: 1, maxLength: 50 }
          description: Partial name; results returned as the caller types.
        - name: level
          in: query
          schema: { type: string, enum: [top, sub] }
        - name: parentId
          in: query
          schema: { type: string }
          description: Restrict to sub-interests of this parent (FR-020).
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InterestPage' }
    post:
      tags: [Interests]
      summary: Create a sub-interest
      description: >
        Serves FR-022. Any signed-in person may create a sub-interest beneath exactly
        one top-level parent. Rate limited per FR-046. Creating a top-level interest
        requires operator authority (FR-021) and is not exposed here.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/InterestCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Interest' }
        '409':
          description: >
            A sufficiently similar sub-interest already exists under this parent
            (FR-023). The response carries the candidates so the caller can join one
            instead of creating a duplicate.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DuplicateInterestCandidates' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
  /interests/{interestId}/posts:
    get:
      tags: [Interests, Posts]
      summary: Posts in an interest, newest first
      description: >
        Serves FR-024, FR-025. For a top-level interest the response includes posts
        published to its sub-interests. Empty results carry an empty-state hint
        (FR-036).
      security: []
      parameters:
        - $ref: '#/components/parameters/InterestId'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        # 004/FR-027, FR-028: `top` REORDERS the set `new` returns. It must not
        # change WHICH posts are returned - SC-009 asserts identical id sets.
        - { name: order, in: query, schema: { type: string, enum: [new, top], default: new } }
        # 004/FR-029: with `q` this is surface 11 of the visibility matrix.
        # Matching happens AFTER filtering, never before.
        - { name: q, in: query, schema: { type: string } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PostPage' }
...
  /people:
    get:
      tags: [People]
      summary: Search people by handle or display name
      description: >
        Serves FR-034, FR-035, FR-036. Excludes people blocked in EITHER direction and any
        person whose account is not active. The block exclusion is server-side and is tested
        through the path a hostile client would take (Constitution III).
      parameters:
        - { name: q, in: query, required: true, schema: { type: string, minLength: 1 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 25, default: 10 } }
      responses:
        '200':
          description: Matching people
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/PublicProfile' } }
        '401': { $ref: '#/components/responses/Unauthorised' }

  # ------------------------------------------------------------------- Saved posts (US5)
  /posts/{postId}/save:
    put:
      tags: [Saved]
      summary: Save a post
      description: Serves FR-037. Refused for a post the caller cannot currently see.
      parameters: [ { name: postId, in: path, required: true, schema: { type: string } } ]
      responses:
        '204': { description: Saved }
        '401': { $ref: '#/components/responses/Unauthorised' }
        '404': { description: No such post, or not visible to the caller }
    delete:
      tags: [Saved]
      summary: Unsave a post
      parameters: [ { name: postId, in: path, required: true, schema: { type: string } } ]
      responses:
        '204': { description: Not saved }
        '401': { $ref: '#/components/responses/Unauthorised' }
  /me/saved:
    get:
      tags: [Saved]
      summary: The signed-in person's saved posts
      description: >
        Serves FR-037, FR-038, FR-039. **Surface 9** of the visibility matrix. A save is a
        bookmark, not a copy: a post the saver may no longer see is absent, never stale
        (SC-013).
      parameters:
        - { name: cursor, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 50, default: 20 } }
      responses:
        '200':
          description: Saved posts the caller can still see, most recently saved first
          content: { application/json: { schema: { $ref: '#/components/schemas/PostPage' } } }
        '401': { $ref: '#/components/responses/Unauthorised' }

  # ----------------------------------------------- Publish and edit (US2) — changed
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Cognito-issued access token.
  parameters:
    Handle:
      name: handle
      in: path
      required: true
      schema: { type: string, pattern: '^[a-z0-9_]{3,30}$' }
    InterestId:
      name: interestId
      in: path
      required: true
      schema: { type: string }
...
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
      description: >
        Opaque continuation token. Paging with this token preserves position
        (FR-035); offsets are deliberately not supported.
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 50, default: 20 }


    # ---------------------------------------------- feature 004
...
  responses:
    Unauthorised:
      description: Missing or invalid credentials
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: Not found
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    ValidationError:
      description: Request failed validation
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    RateLimited:
      description: Rate limit exceeded (FR-046)
      headers:
        Retry-After:
          schema: { type: integer }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }


    # ---------------------------------------------- feature 004
...
  schemas:
    Problem:
      type: object
      description: RFC 9457 problem detail.
      required: [type, title, status]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
    Visibility:
      type: string
      enum: [public, followers, private]
      description: FR-013 to FR-016. Defaults to public on creation.
    ProcessingState:
      type: string
      enum: [pending, processing, ready, failed]
      description: FR-009. Only `ready` posts are visible to anyone but the author.
    SelfProfile:
      type: object
      required: [userId, handle, displayName, notificationPrefs]
      properties:
        userId: { type: string }
        handle: { type: string }
        displayName: { type: string }
        avatarUrl: { type: string, nullable: true }
        bio: { type: string, nullable: true }
        followerCount: { type: integer }
        followingCount: { type: integer }
        interestFollowCount: { type: integer }
        notificationPrefs:
          type: object
          description: >
            FR-049 — one flag per category. `message` is added by 004/FR-031 for
            the conversations that feature introduces. Absent means on, so no
            backfill and no migration.
          properties:
            reaction: { type: boolean }
            comment: { type: boolean }
            follow: { type: boolean }
            message: { type: boolean }
    PublicProfile:
      type: object
      required: [userId, handle, displayName]
      properties:
        userId: { type: string }
        handle: { type: string }
        displayName: { type: string }
        avatarUrl: { type: string, nullable: true }
        bio: { type: string, nullable: true }
        followerCount: { type: integer }
        followingCount: { type: integer }
        topInterests:
          type: array
          description: FR-038 — the interests this person posts to most.
          items: { $ref: '#/components/schemas/InterestRef' }
        viewerIsFollowing: { type: boolean }
    ProfileUpdate:
      type: object
      properties:
        displayName: { type: string, minLength: 1, maxLength: 50 }
        bio: { type: string, maxLength: 300 }
        avatarUploadId: { type: string }
        notificationPrefs:
          type: object
          description: >
            Sent as a PARTIAL patch and merged server-side, so toggling one
            category cannot silently reset the others. `message` is 004/FR-031.
          properties:
            reaction: { type: boolean }
            comment: { type: boolean }
            follow: { type: boolean }
            message: { type: boolean }
    DeletionOutcome:
      type: object
      properties:
        postsScheduledForRemoval: { type: integer }
        commentsToBeAnonymised: { type: integer }
        purgeCompletesBy: { type: string, format: date-time }
    InterestRef:
      type: object
      required: [interestId, name, level]
      properties:
        interestId: { type: string }
        name: { type: string }
        slug: { type: string }
        level: { type: string, enum: [top, sub] }
        parent:
          description: Present for sub-interests, so search results disambiguate (FR-026).
          oneOf:
            - $ref: '#/components/schemas/InterestRef'
            - type: 'null'
    Interest:
      allOf:
        - $ref: '#/components/schemas/InterestRef'
        - type: object
          properties:
            description: { type: string, nullable: true }
            postCount: { type: integer }
            followerCount: { type: integer }
            state: { type: string, enum: [active, merging, merged, retired] }
            viewerIsFollowing: { type: boolean }
...
    InterestCreate:
      type: object
      required: [name, parentId]
      properties:
        name: { type: string, minLength: 2, maxLength: 50 }
        parentId:
          type: string
          description: Must reference a top-level interest — the hierarchy is exactly two levels (FR-020).
        description: { type: string, maxLength: 500 }
        acknowledgedSimilarTo:
          type: array
          description: >
            Interest ids the caller was shown as near-duplicates and chose to create
            alongside anyway (FR-023). Recorded so SC-008 can distinguish an ignored
            warning from one that was never shown.
          items: { type: string }
    DuplicateInterestCandidates:
      type: object
      properties:
        candidates:
          type: array
          items:
            type: object
            properties:
              interest: { $ref: '#/components/schemas/Interest' }
              similarity: { type: number, minimum: 0, maximum: 1 }
    MediaItem:
      type: object
      required: [kind, processingState]
      properties:
        kind: { type: string, enum: [image, video] }
        processingState: { $ref: '#/components/schemas/ProcessingState' }
        width: { type: integer }
        height: { type: integer }
        durationMs: { type: integer, nullable: true }
        posterUrl: { type: string, nullable: true, description: 'Thumbnail shown before playback (FR-009).' }
        renditions:
          type: object
          additionalProperties: { type: string, format: uri }
    Post:
      type: object
      required: [postId, author, interests, visibility, processingState, createdAt]
      properties:
        postId: { type: string }
        author: { $ref: '#/components/schemas/PublicProfile' }
        caption: { type: string, nullable: true, maxLength: 2000 }
        interests:
          type: array
          minItems: 1
          description: FR-006 — never empty.
          items: { $ref: '#/components/schemas/InterestRef' }
        visibility: { $ref: '#/components/schemas/Visibility' }
        processingState: { $ref: '#/components/schemas/ProcessingState' }
        mediaKind: { type: string, enum: [images, video] }
        media:
          type: array
          items: { $ref: '#/components/schemas/MediaItem' }
        reactionCount: { type: integer }
        commentCount: { type: integer }
        viewerHasReacted: { type: boolean }
        place:
          description: >
            004/FR-023. The place the author attached, if any. NEVER derived from
            media metadata (FR-021).
          oneOf:
            - { $ref: '#/components/schemas/PlaceSummary' }
            - { type: 'null' }
        viewerHasSaved: { type: boolean, description: 004/FR-037 }
        createdAt: { type: string, format: date-time }
...
    PageMeta:
      type: object
      properties:
        nextCursor: { type: string, nullable: true }
        emptyStateHint:
          type: string
          nullable: true
          description: >
            FR-036 — which empty state to show, e.g. `no_followed_interests`,
            `interest_has_no_posts`.
    PostPage:
      type: object
      required: [items, page]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Post' }
        page: { $ref: '#/components/schemas/PageMeta' }
    InterestPage:
      type: object
      required: [items, page]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Interest' }
        page: { $ref: '#/components/schemas/PageMeta' }
...
    PlaceCategory:
      type: string
      enum: [restaurant, cafe, bar, shop, venue, outdoor, other]
    PlaceSummary:
      type: object
      required: [placeId, name, category, locality]
      properties:
        placeId: { type: string }
        name: { type: string }
        category: { $ref: '#/components/schemas/PlaceCategory' }
        locality: { type: string }
        postCount: { type: integer }
...